Fixes for intro screen and scale
There was a delay with the intro screen of unknown origin. Turns out it was some primitive weight refreshing I only half implemented causing the screen to only appear once the weight changed. I finished the weight check and that fixed the screen delay. The weight display is improved as a happy coincidence.